LayoutTests/inspector/debugger/content-providers-expected.txt

Tests ContentProvider implementations.


Running: testConcatenatedScriptsContentProvider
  <script>function foo() { return 0; }</script>
                                                              <script>function bar() { return 0; }</script>

            <script>
function baz()
{
  return 0;
}
</script>
